+++ ECWT_train/drawable.h
@@ -10,8 +10,8 @@ class drawable
 {
 	UINT msg = 0;
 public:
-	virtual Eigen::VectorXd points(int) = 0;	//convert to values to be plotted against time
-	virtual void draw(HWND hwnd, Eigen::VectorXd vals)
+	virtual Eigen::VectorXf points(int) = 0;	//convert to values to be plotted against time
+	virtual void draw(HWND hwnd, Eigen::VectorXf vals)
 	{
 		SendMessage(hwnd, this->msg, NULL, reinterpret_cast<LPARAM>(&vals));
 	}
